+++ Examples/TestBed/Tests/MotorsAndLimits.h
@@ -131,7 +131,7 @@ public:
 		float32 torque1 = m_joint1->GetMotorTorque();
 		float32 torque2 = m_joint2->GetMotorTorque();
 		float32 force3 = m_joint3->GetMotorForce();
-		DrawString(5, m_textLine, "Motor Torque = %4.0f, %4.0f : Motor Force = %4.0f", torque1, torque2, force3);
+		DrawString(5, m_textLine, "Motor Torque = %4.0f, %4.0f : Motor Force = %4.0f", (float) torque1, (float) torque2, (float) force3);
 		m_textLine += 15;
 	}
